pyproject.toml
[tool.poetry]
name = "forwarder_bot_deprecated"
version = "0.1.0"
description = ""
authors = ["Carlos Lugones <contact@lugodev.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
